\name{CCIPCA-package}
\alias{CCIPCA-package}
\docType{package}
\title{Candid Covariance-Free Incremental PCA (CCIPCA)}
\description{Incremental (online) PCA without estimation of covariance matrix}
\details{
\tabular{ll}{
Package: \tab CCIPCA\cr
Type: \tab Package\cr
Version: \tab 0.99.0\cr
Date: \tab 2015-01-03\cr
License: \tab Artistic-2.0\cr
}
}
\author{
Koki Tsuyuzaki, Itoshi Nikaido

Maintainer: Koki Tsuyuzaki <k.t.the-answer@hotmail.co.jp>
}
\references{
Weng J. et al. (2003). Candid covariance-free incremental principal component analysis.  \emph{IEEE Transactions on Pattern Analysis and Machine Intelligence}
}
